Share message handling logic between `BuiltInBuildSystemAdapter` and `BuildSystemManager`
